pyproject.toml
[project]
name = "gitreviewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Francisco Filho", email = "francisco@capimgrosso.com" }
]
requires-python = ">=3.13"
dependencies = [
    "gitpython>=3.1.44",
    "google-genai>=1.19.0",
    "ollama>=0.5.1",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "tqdm>=4.67.1",
    "tree-sitter>=0.24.0",
    "tree-sitter-language-pack>=0.7.3",
]

[project.scripts]
gitreviewer = "gitreviewer:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
